cheking emails and Jwt tokens!
This update brought features such as checking the validity of mail to the application (in the future, a check for existence will be added, while this technology is useless, but when the last edits are made to "finish" the project, that is, for the 1.0 implementation, verification via mail and password recovery through it will be added). Jet technology has also been implemented, which I will use in the future to process all customer requests. The algorithm may not be very good in my performance, but I plan to improve it in the future