Sinple list chats
This update was much more difficult for me, I contacted the forum twice, as there were first problems with detecting the JWT token, and then problems at the level of response from the server. At the moment, the server can: register, log in, and after logging in, find all available chats of the authorized user.