mini fix app
there have been no updates for a long time due to burnout from the project. Now I have only got around to fixing the problems that I identified at the end of the work, namely: registration did not take place, since there was a change in the logic of sending and I messed up a little there.